\name{print.di_estimate}
\alias{print.di_estimate}
\alias{print.di_band}
\alias{print.di_set}
\title{Print debiasedInference result objects}
\description{Compact summaries of estimates, confidence bands, and level sets.}
\usage{
\method{print}{di_estimate}(x, ...)

\method{print}{di_band}(x, ...)

\method{print}{di_set}(x, ...)
}
\arguments{
  \item{x}{A result object.}
  \item{...}{Unused.}
}
\value{The input object, invisibly.}
